PreparingfortheOlympiad

+++ #2050B/Transfusion.cpp
@@ -11,14 +11,14 @@ int main()
     {
         int n;
         cin >> n;
-        vector<int> arr(n);
+        vector<long long> arr(n);
         for (int i = 0; i < n; i++)
         {
             cin >> arr[i];
         }
 
-        int odd_sum = 0;
-        int even_sum = 0;
+        long long odd_sum = 0;
+        long long even_sum = 0;
 
         for (int i = 0; i < n; i++)
         {
@@ -32,8 +32,8 @@ int main()
             }
         }
 
-        int odd_places = n / 2;
-        int even_places = (n + 1) / 2;
+        long long odd_places = n / 2;
+        long long even_places = (n + 1) / 2;
 
         if (odd_sum % odd_places == 0 && even_sum % even_places == 0 && odd_sum / odd_places == even_sum / even_places)
         {
